+++ CHANGELOG.md
@@ -7,6 +7,11 @@ Bug-fixes within the same version aren't needed
 
 -->
 
-### 0.0.0
+### 0.0.2
 
-* TBD
+* Templates: dotnet (windows only)
+
+### 0.0.1
+
+* Commands: init, refresh
+* Templates: nodejs, go, rust, python, jupyterlab